{"types":"C","runs":6}

+++ problem15.java
@@ -5,7 +5,7 @@ public class problem15 extends ConsoleProgram {
 		int n1 = readInt();
 		int n2 = readInt();
 		int total = n1 + n2;
-		int arithmeticMiddle = total/2;
+		double arithmeticMiddle = total/2;
 		println(arithmeticMiddle);
 	}
 }
